requirements.txt
tensorflow==1.15.2
tensorflow-gpu==1.15.2
gym-retro==0.8.0
stable-baselines==2.10.0
